# cmd/api/docs/swagger.yaml
definitions:
...
  models.Product:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        type: number
      size:
        type: integer
      stock:
        type: integer
    required:
    - category_id
    - description
    - name
    - price
    - size
    - stock
    type: object
...
  models.ProductUpdate:
    properties:
      product_id:
        type: integer
      stock:
        type: integer
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Zhooze is an E-commerce platform to purchase and sell shoes
  title: Go + Gin Zhooze E-Commerce API
  version: 1.0.0
paths:
...
  /admin/products:
    delete:
      consumes:
      - application/json
      description: Delete a product from the admin side
      parameters:
      - description: product id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete product
      tags:
      - Admin Product Management
    get:
      consumes:
      - application/json
      description: Retrieve all product Details
      parameters:
      - description: Page number
        in: query
        name: page
        type: string
      - description: Page Count
        in: query
        name: count
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Products Details
      tags:
      - Admin Product Management
    post:
      consumes:
      - application/json
      description: Add product from admin side
      parameters:
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Products
      tags:
      - Admin Product Management
    put:
      consumes:
      - application/json
      description: Update quantity of already existing product
      parameters:
      - description: Product details
        in: body
        name: productUpdate
        required: true
        schema:
          $ref: '#/definitions/models.ProductUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update Products quantity
      tags:
      - Admin Product Management
  /admin/products/upload-image:
    post:
      consumes:
      - application/json
      description: Add product Product from admin side
      parameters:
      - description: Product_id
        in: query
        name: product_id
        required: true
        type: integer
      - description: Image file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Product Image
      tags:
      - Admin Product Management
...
  /user/products:
    get:
      consumes:
      - application/json
      description: Retrieve all product Details with pagination to users
      parameters:
      - description: Page number
        in: query
        name: page
        type: string
      - description: Page Count
        in: query
        name: count
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Products Details to users
      tags:
      - User Product
  /user/products/filter:
    post:
      consumes:
      - application/json
      description: Show all the Products belonging to a specified category
      parameters:
      - description: Category IDs and quantities
        in: body
        name: data
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Show Products of specified category
      tags:
      - User Product
  /user/products/image:
    get:
      consumes:
      - application/json
      description: Retrieve product images
      parameters:
      - description: product_id
        in: query
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Products Details to users
      tags:
      - User Product
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"